requirements.txt
asgiref==3.7.2
dj-database-url
Django
django-environ
djangorestframework
gunicorn
packaging==23.1
Pillow
psycopg2
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
django-cors-headers
whitenoise